requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.0
numpy>=1.23
opencv-python-headless>=4.8
Pillow>=10.0
onnxruntime==1.15.1            
torch>=2.0.0
torchvision>=0.15.0
grad-cam==1.4.6                
ultralytics==8.3.203           
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.1.0
